# bauth_server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bauth",
    "description": "Headless authentication server. App routes answer errors with `ErrorBody` and a stable `code`; `/oauth/*` follows RFC 6749.",
    "contact": {
      "name": "Egor Berezovskiy"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/email-change/confirm": {
      "post": {
        "tags": [
          "Me"
        ],
        "summary": "Called by the app page the email change link opens (`email_change_url`): moves the account to\nthe new address (`POST /me/email`), which the link proves the user owns.",
        "operationId": "confirm_email_change",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmEmailChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Email change applied"
          },
          "400": {
            "description": "`invalid_request`, `invalid_token`, `account_disabled`, `email_taken`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/me": {
      "get": {
        "tags": [
          "Me"
        ],
        "operationId": "get_me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized`: refresh the access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Me"
        ],
        "summary": "Deletes the account right away, with every session.\nAccess tokens already given to APIs stay valid until they expire (15 min).",
        "operationId": "delete_account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Account deleted"
          },
          "400": {
            "description": "`invalid_request`, `invalid_code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized`: refresh the access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/me/confirmation": {
      "post": {
        "tags": [
          "Me"
        ],
        "summary": "Emails a 6-digit code confirming `action`, to send back as `code` on that route: a stolen\naccess token alone can't change the email or delete the account. The code only works for this\naction, from this session, for 5 wrong attempts; a new one disables the previous.",
        "operationId": "request_confirmation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Code emailed to the account's address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmationResponse"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized`: refresh the access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/me/email": {
      "post": {
        "tags": [
          "Me"
        ],
        "summary": "Sends a confirmation link to the new address. The account keeps its current address\nuntil the link is used, on `POST /email-change/confirm`. Confirmed by a code from\n`POST /me/confirmation`.",
        "operationId": "change_email",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Confirmation link sent to the new address, unless it is taken: the answer is the same, and a `code` is spent either way",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeEmailResponse"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request`, `invalid_email`, `invalid_code`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized`: refresh the access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/me/sessions": {
      "get": {
        "tags": [
          "Me"
        ],
        "operationId": "list_sessions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized`: refresh the access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/me/sessions/{session_id}": {
      "delete": {
        "tags": [
          "Me"
        ],
        "summary": "Logs out one device. Revoking the current session works too.",
        "operationId": "revoke_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "From `GET /me/sessions`",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Session revoked"
          },
          "401": {
            "description": "`unauthorized`: refresh the access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "`not_found`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangeEmailRequest": {
        "type": "object",
        "required": [
          "code",
          "new_email"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "6-digit code from `POST /me/confirmation` with `action: \"change_email\"`."
          },
          "new_email": {
            "type": "string"
          }
        }
      },
      "ChangeEmailResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ChangeEmailStatus"
          }
        }
      },
      "ChangeEmailStatus": {
        "type": "string",
        "enum": [
          "confirmation_sent"
        ]
      },
      "ConfirmEmailChangeRequest": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
...
      "ConfirmationAction": {
        "type": "string",
        "description": "What a code confirms: sensitive changes need proof of the mailbox, not just an access token.",
        "enum": [
          "change_email",
          "delete_account"
        ]
      },
      "ConfirmationRequest": {
        "type": "object",
        "required": [
          "action"
        ],
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ConfirmationAction"
          }
        }
      },
      "ConfirmationResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ConfirmationStatus"
          }
        }
      },
      "ConfirmationStatus": {
        "type": "string",
        "enum": [
          "confirmation_sent"
        ]
      },
...
      "DeleteAccountRequest": {
        "type": "object",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "6-digit code from `POST /me/confirmation` with `action: \"delete_account\"`."
          }
        }
      },
      "ErrorBody": {
        "type": "object",
        "description": "Error of every app route. Apps translate `code`, which is stable; `message` is for developers.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "invalid_code"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "MeResponse": {
        "type": "object",
        "required": [
          "id",
          "email",
          "email_verified",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": "string"
          },
          "email_verified": {
            "type": "boolean"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "SessionResponse": {
        "type": "object",
        "required": [
          "id",
          "client_id",
          "created_at",
          "last_used_at",
          "expires_at",
          "current"
        ],
        "properties": {
          "client_id": {
            "type": "string"
          },
          "client_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Display name from bauth.toml, `None` if the client was removed since."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "current": {
            "type": "boolean",
            "description": "The session this request's access token belongs to."
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "last_used_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token from `/oauth/token`"
      }
    }
  }
}